mlir: Use mlir's clang-format in libsolidity/codegen/mlir. NFC
The mlir directory is expected to have more code in the mlir namespace
than in the solc namespaces.

Last I checked, unlike the source in the mlir repo, clang-format is not
strictly applied to the existing solc source.

Also, sticking to the mlir style will make things easier if we decide to
integrate our dialects in the mlir repo (assuming that it makes sense
and the licenses allow us to do so).