[mlir][EraVM] Generate (un)signed types in the solidity ast lowering by using custom arith ops in the sol dialect
This is a big one! We completely drop signless types and the arith
dialect from the solidity ast lowering. We still need to address some
more signless cases like `msg.sender`. With this change, we should keep
signless types in yul and stage-2 lowering. The rationale for signed
types is documented in docs/internal/mlir.rst.